Merge in develop, resolve conflics, adjustements to AMP_Travel_Theme().
In that class, make all functions public.
Change __construct() to init(),
and move the requiring of class-amp-travel-cpt.php
to AMP_Travel_Theme::includes().
Thanks to David for creating that file.
This keeps all of the logic of the plugin files self-sufficient.
functions.php wil only have to require class-amp-travel-theme.php.
And call amp_travel_theme()->init().