Change featured/index.js save() to return null.
Before, there was an error in Gutenberg,
as renderStaticFeaturedBlock() wasn't defined.
This PR removed that function.
So change save() to return null,
as this looks to be a dynamic block.
AMP_Travel_Blocks::render_block_travel_featured()
renders the output on the front-end.
Also, compile js/editor-blocks.js, using:
npm run build.