Change 'Hero' block placeholders into default values.
The placeholders in React can seem deceptive.
They appear in the editor,
but if you save the block without editing it,
they don't appear on the front-end.
So change them to default values.
So they will initially appear in the editor,
and they'll appear on the front-end as defaults.
Also, make the 'key' values unique.
There have been React warnings for duplicate keys,
though maybe not specifically for this:
Warning: Encountered two children with the same key.