Hide AMP plugin validator notices for the screencast.
The home page has a notice for the entire page:
"Too much CSS is enqueued and so seemingly irrelevant rules have been removed."
This is a valid notice,
but it'd be better to not have it in the screencast.
Also, the hero block on the top has a notice
from the amp-list not using https for its src.
This new style rule hides both notices.
We can revert this later.
And the plugin will have a way to ignore them.